Only print user owned business in user business page.

+++ Support_Your_Locals/Controllers/BusinessController.cs
@@ -59,7 +59,7 @@ namespace Support_Your_Locals.Controllers
                 // Exception here
                 Header = businessRegisterModel.Header,
                 Description = businessRegisterModel.Description,
-                UserID = Int32.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier).Value),
+                UserID = long.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier).Value),
                 PhoneNumber = businessRegisterModel.PhoneNumber,
                 Latitude = businessRegisterModel.Latitude,
                 Longitude = businessRegisterModel.Longitude,
